requirements.txt
dbt-core
dbt-duckdb
ghp-import
polars